Rename AgentService to ConversationRuntime, extract ensureConversation and resolveToolGroupsFromRegistry
The "AgentService" was a god-service wrapping the conversation engine.
This renames it to ConversationRuntime to reflect its actual role, extracts
ensureConversation into conversation/ensure.ts (standalone DB upsert) and
resolveToolGroupsFromRegistry into conversation/tools.ts (plugin tool
resolution). Deprecated aliases preserve backwards compatibility.
