+++ apps/api/src/scripts/test-telegram-flow.ts
@@ -8,7 +8,7 @@
  * Usage: doppler run -- bun run scripts/test-telegram-flow.ts
  */
 
-import { AgentService, ModelServiceLive, makeAgentServiceLive } from "@amby/agent"
+import { ConversationRuntime, ModelServiceLive, makeConversationRuntimeLive } from "@amby/agent"
 import { AuthServiceLive } from "@amby/auth"
 import { BrowserServiceDisabledLive } from "@amby/browser/local"
 import { SandboxServiceLive, TaskSupervisorLive } from "@amby/computer"
@@ -182,9 +182,9 @@ async function main() {
 	try {
 		maybeConversationId = await runtime.runPromise(
 			Effect.gen(function* () {
-				const agent = yield* AgentService
+				const agent = yield* ConversationRuntime
 				return yield* agent.ensureConversation("telegram", String(SIMULATED_CHAT_ID))
-			}).pipe(Effect.provide(makeAgentServiceLive(userId))),
+			}).pipe(Effect.provide(makeConversationRuntimeLive(userId))),
 		)
 
 		record("Conversation created/found", true)
@@ -204,9 +204,9 @@ async function main() {
 	try {
 		const convId2 = await runtime.runPromise(
 			Effect.gen(function* () {
-				const agent = yield* AgentService
+				const agent = yield* ConversationRuntime
 				return yield* agent.ensureConversation("telegram", String(SIMULATED_CHAT_ID))
-			}).pipe(Effect.provide(makeAgentServiceLive(userId))),
+			}).pipe(Effect.provide(makeConversationRuntimeLive(userId))),
 		)
 
 		if (convId2 === conversationId) {
@@ -223,9 +223,9 @@ async function main() {
 	try {
 		const result = await runtime.runPromise(
 			Effect.gen(function* () {
-				const agent = yield* AgentService
+				const agent = yield* ConversationRuntime
 				return yield* agent.handleMessage(conversationId, "Hello! How are you?")
-			}).pipe(Effect.provide(makeAgentServiceLive(userId))),
+			}).pipe(Effect.provide(makeConversationRuntimeLive(userId))),
 		)
 
 		if (result.userResponse.text.trim().length > 0) {
@@ -244,12 +244,12 @@ async function main() {
 	try {
 		const result = await runtime.runPromise(
 			Effect.gen(function* () {
-				const agent = yield* AgentService
+				const agent = yield* ConversationRuntime
 				return yield* agent.handleMessage(
 					conversationId,
 					"Remember this: my favorite programming language is TypeScript",
 				)
-			}).pipe(Effect.provide(makeAgentServiceLive(userId))),
+			}).pipe(Effect.provide(makeConversationRuntimeLive(userId))),
 		)
 
 		record("Memory save", true)
@@ -264,12 +264,12 @@ async function main() {
 	try {
 		const result = await runtime.runPromise(
 			Effect.gen(function* () {
-				const agent = yield* AgentService
+				const agent = yield* ConversationRuntime
 				return yield* agent.handleMessage(
 					conversationId,
 					"Investigate what day of the week March 25, 2026 falls on",
 				)
-			}).pipe(Effect.provide(makeAgentServiceLive(userId))),
+			}).pipe(Effect.provide(makeConversationRuntimeLive(userId))),
 		)
 
 		record("Research request", true)
@@ -284,9 +284,9 @@ async function main() {
 	try {
 		const result = await runtime.runPromise(
 			Effect.gen(function* () {
-				const agent = yield* AgentService
+				const agent = yield* ConversationRuntime
 				return yield* agent.handleMessage(conversationId, "Set my timezone to America/Los_Angeles")
-			}).pipe(Effect.provide(makeAgentServiceLive(userId))),
+			}).pipe(Effect.provide(makeConversationRuntimeLive(userId))),
 		)
 
 		record("Timezone setting", true)
