+++ packages/agent/src/index.ts
@@ -6,6 +6,8 @@ export {
 	handleTurn,
 	type TurnRequest,
 } from "./conversation/engine"
+export { ensureConversation } from "./conversation/ensure"
+export { resolveToolGroupsFromRegistry } from "./conversation/tools"
 export * from "./errors"
 export * from "./models"
 export * from "./router"
